requirements.txt
asgiref==3.5.0
Django==3.2.12
django-crispy-forms==1.14.0
Pillow==9.0.1
pytz==2021.3
sqlparse==0.4.2
